backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
pydantic
passlib[bcrypt]
email-validator
openai>=1.0.0
python-multipart>=0.0.6
requests 
python-dotenv
httpx
pytesseract
